generator.py
from fpdf.template import Template
from fpdf import FPDF
from orm import Pack, PackSubTask, Customer
import os
from datetime import date as dt
import hashlib as hb
import time
from setup import DOCS_PATH
