requirements.txt
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==36.0.2
cssselect==1.1.0
cssutils==2.4.0
dateparser==1.1.1
DateTimeRange==1.2.0
defusedxml==0.7.1
dkimpy==1.0.5
dnspython==2.2.1
Flask==2.1.2
fpdf2==2.5.5
google-api-core==2.8.0
google-api-python-client==2.49.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.1
keyring==23.5.0
lxml==4.8.0
MarkupSafe==2.1.1
mbstrdecoder==1.1.0
oauthlib==3.2.0
packaging==21.3
pdfkit==1.0.0
peewee==3.14.10
Pillow==9.2.0
premailer==3.10.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
SecretStorage==3.3.2
six==1.16.0
tabulate==0.8.9
toml==0.10.2
typepy==1.3.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
yagmail==0.15.277
zipp==3.8.0
